requirements.txt
Django==4.2.1
Pillow==9.5.0
django-debug-toolbar==3.6.0
django-mptt==0.14.0
beautifulsoup4==4.12.2
requests==2.31.0